+++ Headers/Destructible.h
@@ -2,13 +2,13 @@
 //actor can take damage and be destroyed
 class Destructible : public Persistent {
 public:
-	float maxHp; // maximum Health points
-	float hp; // current health points
-	float defense; //hit point deflected
+	double maxHp; // maximum Health points
+	double hp; // current health points
+	double defense; //hit point deflected
 	std::string corpseName; // the actors name once dead
 	int xp; //xp for killing monster or player xp
 
-	Destructible(float maxHp, float defense, std::string_view corpseName, int xp);
+	Destructible(double maxHp, double defense, std::string_view corpseName, int xp);
 	
 	//return if actor has died
 	inline bool isDead() {
@@ -16,8 +16,8 @@ public:
 	}
 
 
-	float takeDamage(Actor* owner, float damage); //if actor is attacked work out how much hp is lost
-	float heal(int amount); //restore hp to actor
+	double takeDamage(Actor* owner, double damage); //if actor is attacked work out how much hp is lost
+	double heal(int amount); //restore hp to actor
 
 	virtual void die(Actor* owner); //what actor does when it dies
 
@@ -36,7 +36,7 @@ protected:
 //monster destructible dies and turns into corpse
 class MonsterDestructible : public Destructible {
 public:
-	MonsterDestructible(float maxHp, float defense, std::string_view corpseName, int xp);
+	MonsterDestructible(double maxHp, double defense, std::string_view corpseName, int xp);
 	void die(Actor* owner);//turns actor into corpse
 	void save(TCODZip& zip);//save destructible data to zip
 	
@@ -45,7 +45,7 @@ public:
 //player dies and game ends
 class PlayerDestructible : public Destructible {
 public:
-	PlayerDestructible(float maxHp, float defense, std::string_view corpseName, int xp);
+	PlayerDestructible(double maxHp, double defense, std::string_view corpseName, int xp);
 	void die(Actor* owner);//ends game
 	void save(TCODZip& zip);//save destructible data to zip
